### 2023-01-20 #4994 @Kevin-Mattheus-Moerman
@uelegre please respond on where the paper is. I believe it is this file: https://github.com/BraiNEdarwin/brains-py/blob/master/docs/sample.md, but it should be named `paper.md`, also, it looks like the paper is not formatted according to [our requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), for instance it should contain a `Statement of need` section. If you are interested in pursuing review with JOSS we require that you address these and respond to queries raised here.
If we do not hear from you within a week I will assume you are no longer interesting in pursuing review with JOSS and we will reject this submission.
