### 2023-01-30 #4994 @ualegre
I was not expecting communications to be over Github. I am still interested in pursuing a review with JOSS. I will address these requests ASAP.
