### 2023-01-31 b8372c3
Fixes to submit according to JOSS requirements
