+++ .github/workflows/draft-pdf.yml
@@ -14,7 +14,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: docs/sample.md
+          paper-path: docs/paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
