### 2023-01-31 #4994 @ualegre
@Kevin-Mattheus-Moerman I have now added the Statement of need section and updated the format of paper.md file according to the requirements. This file can be found in the docs folder.
