### 2023-02-06 4082ff6
Updated the yml for the joss compilation
