+++ .github/workflows/draft-pdf.yml
@@ -6,15 +6,13 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
-      - name: Open Journals PDF Generator
-	uses: openjournals/openjournals-draft-action@v.1.0
+        uses: actions/checkout@v3
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: docs/paper.md
+          paper-path: paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -22,4 +20,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: docs/paper.pdf
+          path: paper.pdf
