### 2023-02-06 #4994 @ualegre
@Kevin-Mattheus-Moerman there seemed to be some issues related to the way in which I had uploaded the .yml file for specifying the action of generating the pdf. It now works on the workflow on my repository so if we generate it again it should work now.
