### 2023-02-10 f5c5efb
Cleaned the bib of the paper
