+++ docs/paper.bib
@@ -1,3 +1,38 @@
+@article{kaspar2021rise,
+  title={The rise of intelligent matter},
+  author={Kaspar, C and Ravoo, BJ and van der Wiel, Wilfred G and Wegner, SV and Pernice, WHP},
+  journal={Nature},
+  volume={594},
+  number={7863},
+  pages={345--355},
+  year={2021},
+  publisher={Nature Publishing Group}
+}
+
+@article{chen2020classification,
+  title={Classification with a disordered dopant-atom network in silicon},
+  author={Chen, Tao and van Gelder, Jeroen and van de Ven, Bram and Amitonov, Sergey V and de Wilde, Bram and Euler, Hans-Christian Ruiz and Broersma, Hajo and Bobbert, Peter A and Zwanenburg, Floris A and van der Wiel, Wilfred G},
+  journal={Nature},
+  volume={577},
+  number={7790},
+  pages={341--345},
+  year={2020},
+  publisher={Nature Publishing Group}
+}
+
+@article{ruiz2021dopant,
+	doi = {10.1088/2634-4386/ac1a7f},
+	year = 2021,
+	month = {sep},
+	publisher = {{IOP} Publishing},
+	volume = {1},
+	number = {2},
+	pages = {024002},
+	author = {Hans-Christian Ruiz-Euler and Unai Alegre-Ibarra and Bram van de Ven and Hajo Broersma and Peter A Bobbert and Wilfred G van der Wiel},
+	title = {Dopant network processing units: towards efficient neural network emulators with high-capacity nanoelectronic nodes},
+	journal = {Neuromorphic Computing and Engineering},
+}
+
 @incollection{sivanandam2008genetic,
   title={Genetic algorithms},
   author={Sivanandam, SN and Deepa, SN},
@@ -7,6 +42,16 @@
   publisher={Springer}
 }
 
+@article{ruiz2020deep,
+  title={A deep-learning approach to realizing functionality in nanoelectronic devices},
+  author={Ruiz Euler, Hans-Christian and Boon, Marcus N and Wildeboer, Jochem T and van de Ven, Bram and Chen, Tao and Broersma, Hajo and Bobbert, Peter A and van der Wiel, Wilfred G},
+  journal={Nature nanotechnology},
+  volume={15},
+  number={12},
+  pages={992--998},
+  year={2020},
+  publisher={Nature Publishing Group}
+}
 
 @inproceedings{dogo2018comparative,
   title={A comparative analysis of gradient descent-based optimization algorithms on convolutional neural networks},
@@ -31,1026 +76,4 @@
   journal={Advances in neural information processing systems},
   volume={32},
   year={2019}
-}
-
-@article{boon2021gradient,
-  title={Gradient descent in materio},
-  author={Boon, Marcus N and Euler, Hans-Christian Ruiz and Chen, Tao and van de Ven, Bram and Ibarra, Unai Alegre and Bobbert, Peter A and van der Wiel, Wilfred G},
-  journal={arXiv preprint arXiv:2105.11233},
-  year={2021}
-}
-
-@article{kaspar2021rise,
-  title={The rise of intelligent matter},
-  author={Kaspar, C and Ravoo, BJ and van der Wiel, Wilfred G and Wegner, SV and Pernice, WHP},
-  journal={Nature},
-  volume={594},
-  number={7863},
-  pages={345--355},
-  year={2021},
-  publisher={Nature Publishing Group}
-}
-
-
-@article{lecun1998gradient,
-  title={Gradient-based learning applied to document recognition},
-  author={LeCun, Yann and Bottou, L{\'e}on and Bengio, Yoshua and Haffner, Patrick},
-  journal={Proceedings of the IEEE},
[…]
