### 2023-02-10 e7e4034
Added missing blumer paper to bibliography
