+++ docs/paper.bib
@@ -77,3 +77,14 @@
   volume={32},
   year={2019}
 }
+
+@article{blumer1989learnability,
+  title={Learnability and the Vapnik-Chervonenkis dimension},
+  author={Blumer, Anselm and Ehrenfeucht, Andrzej and Haussler, David and Warmuth, Manfred K},
+  journal={Journal of the ACM (JACM)},
+  volume={36},
+  number={4},
+  pages={929--965},
+  year={1989},
+  publisher={ACM New York, NY, USA}
+}
