### 2023-02-10 53b1af2
Tried removing related publications
