### 2023-02-10 ed3c944
Tested to remove references to images
