### 2023-02-10 7b10493
Tried to remove a couple of sections
