### 2023-02-10 66aa65f
Added citation of ruiz2021dopant
