### 2023-02-10 1d6aaf9
Added back projects and related publications sections
