+++ docs/paper.md
@@ -73,6 +73,17 @@ The breakthroughs of deep learning come along with high energy costs, related to
 
 The framework is composed of two main packages, brains-py (core) and brainspy-smg (surrogate model generator). The former package contains the whole structure for processors, managing the drivers for National Instruments devices that are connected to the DNPUs, and a set of utils functions that can be used for managing the signals, custom loss/fitness functions, linear transformations and i/o file management. The latter package contains the libraries required to prepare DNPU devices for data gathering (multiple and single IV curves), to gather information from multi-terminal devices in an efficient way [@ruiz2020deep], as well as training support for surrogate models, and consistency checks.
 
+# Finding functionality on a single DNPU circuit design
+
+There are two main flavours in which single DNPUs can be trained: *Evolution in Matter* or *Surrogate model based gradient descent*. The *evolution in matter* approach performs a stochastic search for suitable control voltages for a single DNPU unit, measured in hardware, for a given supervised task [@sivanandam2008genetic]. It provides a quicker exploration (directly on hardware) of the usefulness of newly manufactured single DNPUs. Typically, common classification benchmarks are used, such as solving non-linear boolean gates [@chen2020classification] or measuring the Vapnik-Chervonenkis dimension [@ruiz2021dopant]. On the other hand, the *surrogate model* approach [@ruiz2020deep] is better suited for studying the scaling of DNPU hardware. The process is as follows:
+
+1. Generate a surrogate model: For this, the multi-dimensional input-output data of the device is densely sampled. The input consists of sinusoidal or triangular modulation functions, chosen in such a way that the ratios of all frequency combinations are irrational, guaranteeing a densely covered input-output space without repetitions. A DNN is trained for representing the function of this input-output data.
+
+2. Train for DNPU functionality: The weights of the trained DNN are frozen, and the control voltages are declared as learnable parameters of the surrogate model of the DNPU. The training for DNPU functionality is supported by the brains-py framework but is also possible to use customised user-created gradient descent [@dogo2018comparative] algorithms using PyTorch.
+
+3. Validate on hardware: Once satisfactory control voltages for a task are found, brains-py supports seamlessly validating them on hardware, without having to modify the code of the model, by simply changing the model to hardware-evaluation mode.
+
+
 # Finding functionality on multi-DNPU circuit design
 
[…]
