### 2023-02-10 d2c2974
Removed labels inside images
