### 2023-02-10 #4994 @ualegre
@Kevin-Mattheus-Moerman It should be fixed now. I was confident it would pass before as it was passing for my own repository. I checked on the logs of my own github actions and it appears that, although passing, it did not like how the images were labelled. I just removed the labels now.
