### 2023-02-11 #4994 @Kevin-Mattheus-Moerman
@ualegre below are some points on the paper that require your attention:

- [ ] Please rename `Abstract` to `Summary`, as this is customary for this journal.
- [ ] Please check why the heading `#Framework description` does not render properly.
- [ ] Remove the "Related publications" section and instead work such citations into the relevant text/sections discussing the background and state of the art. This way these papers will just be listed with the rest of the bibliography.

Once you've worked on the above you can call `@editorialbot generate pdf` here to update the paper. Similarly you can call `@editorialbot check references` to do a reference/DOI check. I'll trigger that check now too so you may work on any potentially missing/invalid DOIs.
