### 2023-02-13 5d52bc2
Added missing DOIs to bibliography
