+++ docs/paper.bib
@@ -17,6 +17,7 @@
   number={7790},
   pages={341--345},
   year={2020},
+  doi = {10.1038/s41586-019-1901-0},
   publisher={Nature Publishing Group}
 }
 
@@ -44,6 +45,7 @@
 
 @article{ruiz2020deep,
   title={A deep-learning approach to realizing functionality in nanoelectronic devices},
+  doi = {10.1038/s41565-020-00779-y},
   author={Ruiz Euler, Hans-Christian and Boon, Marcus N and Wildeboer, Jochem T and van de Ven, Bram and Chen, Tao and Broersma, Hajo and Bobbert, Peter A and van der Wiel, Wilfred G},
   journal={Nature nanotechnology},
   volume={15},
@@ -55,6 +57,7 @@
 
 @inproceedings{dogo2018comparative,
   title={A comparative analysis of gradient descent-based optimization algorithms on convolutional neural networks},
+  doi = {10.1109/CTEMS.2018.8769211},
   author={Dogo, EM and Afolabi, OJ and Nwulu, NI and Twala, Bhekisipho and Aigbavboa, CO},
   booktitle={2018 international conference on computational techniques, electronics and mechanical systems (CTEMS)},
   pages={92--99},
