### 2023-02-13 6afb326
Deleted related papers section. Added missing references.
