+++ docs/paper.bib
@@ -91,3 +91,27 @@
   year={1989},
   publisher={ACM New York, NY, USA}
 }
+
+@article{tertilt2022hopping,
+  title={Hopping-transport mechanism for reconfigurable logic in disordered dopant networks},
+  doi={10.1103/PhysRevApplied.17.064025},
+  author={Tertilt, Henri and Bakker, Jesse and Becker, Marlon and de Wilde, Bram and Klanberg, Indrek and Geurts, Bernard J and van der Wiel, Wilfred G and Heuer, Andreas and Bobbert, Peter A},
+  journal={Physical Review Applied},
+  volume={17},
+  number={6},
+  pages={064025},
+  year={2022},
+  publisher={APS}
+}
+
+@article{chen20211,
+  title={1/f noise and machine intelligence in a nonlinear dopant atom network},
+  doi={10.1002/smsc.202000014},
+  author={Chen, Tao and Bobbert, Peter A and van der Wiel, Wilfred G},
+  journal={Small Science},
+  volume={1},
+  number={3},
+  pages={2000014},
+  year={2021},
+  publisher={Wiley Online Library}
+}
