### 2023-02-25 #4994 @arfon
👋 @aparna-aketi – would you be willing to review this submission for JOSS? The submission under consideration is brains-py: A framework to support research on energy-efficient unconventional hardware for machine learning (https://github.com/braiNEdarwin/brains-py).

Many thanks
Arfon
