### 2023-05-28 #4994 @arfon
@benedictjones – would you be willing to review this submission for JOSS? The submission under consideration is brains-py: A framework to support research on energy-efficient unconventional hardware for machine learning (https://github.com/braiNEdarwin/brains-py).

Many thanks
Arfon
