### 2023-06-20 #5573 @editorialbot
[REVIEW]: brains-py: A framework to support research on energy-efficient unconventional hardware for machine learning
**Submitting author:** @ualegre (<a href="http://orcid.org/0000-0001-5957-7945">Unai Alegre-Ibarra</a>)
**Repository:** https://github.com/braiNEdarwin/brains-py
**Branch with paper.md** (empty if default branch): master
**Version:** 1.0.2
