### 2023-07-05 BraiNEdarwin/brains-py#66 @wob86
Installation guide not correct - JOSS review
Step 3 says to  run
pip install brains-py
This doesnt work as in setup.py it is refereced as brainspy so the command
pip install brainspy works
Suggest either editing setup file or wiki to correct this

JOSS review
https://github.com/openjournals/joss-reviews/issues/5573
