### 2023-07-05 BraiNEdarwin/brains-py#67 @wob86
Instructions on main page - review
In the main page in section 1 Instructions, you have a very nicely set out overview of the different areas and links to each part, unfortunatly they all link back to the main page of the wiki where all of the same links are presented.  Could these be altered to point to the specific page.

JOSS review
https://github.com/openjournals/joss-reviews/issues/5573
