### 2023-07-18 #5573 @wob86
Paper looks good as does supporting documentation in repo and gives a good introduction to the software and explanation of its need and usage.
I have two main issues one with readability of help sections which I have raised as issue number 67 on the repo and the installation guide not being correct which is why I have left the installatioon step unchecked in the checklist, this is issue. 66
