### 2023-08-09 BraiNEdarwin/brains-py#67 @ualegre
I fixed the links on the main page.
