### 2023-09-10 #5573 @arfon
@ljvmiranda921 @sisco0 – would you be willing to review this submission for JOSS? The submission under consideration is brains-py: A framework to support research on energy-efficient unconventional hardware for machine learning (https://github.com/braiNEdarwin/brains-py).

Many thanks
Arfon
