### 2023-09-22 #5573 @sisco0
The current review is complete.
The whole set of points have been reviewed and verified.
No expected modification to perform over the paper nor the Git repository.
