### 2023-10-05 1d3ca41
Minor change on the simulation processor test
