### 2023-10-05 f69c401
Update README.md
Minor fix on the first reference
