### 2023-10-08 #5573 @arfon
Looks like the DOI is being slow to resolve. I'll keep this issue open until it's registered and resolving.
