### 2023-10-09 #5573 @arfon
Looks like the DOI is resolving now ⚡
